[SPARK-23693][SQL] Refactor: Extract common functions for UUID SQL expressions to an abstract superclass
