+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/expressions/RandomBasedUuid.scala
@@ -19,31 +19,15 @@ package org.apache.spark.sql.catalyst.expressions
 
 import java.util.concurrent.ThreadLocalRandom
 
-import com.fasterxml.uuid.{Generators, NoArgGenerator}
+import com.fasterxml.uuid.Generators
 
-import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions.codegen.{CodegenContext, ExprCode, FalseLiteral}
-import org.apache.spark.sql.types.{DataType, StringType}
-import org.apache.spark.unsafe.types.UTF8String
 
-case class RandomBasedUuid() extends LeafExpression with Stateful {
-
-  @transient private var generator: NoArgGenerator = _
+case class RandomBasedUuid() extends UuidExpression {
 
   override protected def initializeInternal(partitionIndex: Int): Unit = {
     generator = Generators.randomBasedGenerator(ThreadLocalRandom.current())
   }
-
-  override protected def evalInternal(input: InternalRow): Any = {
-    val uuid = generator
-      .generate()
-      .toString
-    UTF8String.fromString(uuid)
-  }
-
-  override def nullable: Boolean = false
-
-  override def dataType: DataType = StringType
 
   override def freshCopy(): RandomBasedUuid = RandomBasedUuid()
