+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/expressions/UuidExpression.scala
@@ -1,0 +1,41 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.spark.sql.catalyst.expressions
+
+import com.fasterxml.uuid.NoArgGenerator
+
+import org.apache.spark.sql.catalyst.InternalRow
+import org.apache.spark.sql.types.{DataType, StringType}
+import org.apache.spark.unsafe.types.UTF8String
+
+abstract class UuidExpression extends LeafExpression with Stateful {
+
+  @transient protected var generator: NoArgGenerator = _
+
+  override protected def evalInternal(input: InternalRow): Any = {
+    val uuid = generator
+      .generate()
+      .toString
+    UTF8String.fromString(uuid)
+  }
+
+  override def nullable: Boolean = false
+
+  override def dataType: DataType = StringType
+
+}
